requirements.txt
# app requirements
pandas
numpy
scipy
matplotlib

#report requirements
reportlab
Pillow
openai==0.28
boto3